backend/pyproject.toml
[project]
name = "kinescope-tms-backend"
version = "0.1.0"
description = "Kinescope TMS — backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.32,<1.0",
    "pydantic[email]>=2.9,<3.0",
    "pydantic-settings>=2.6,<3.0",
    "sqlalchemy[asyncio]>=2.0.36,<3.0",
    "asyncpg>=0.29,<1.0",
    "alembic>=1.14,<2.0",
    "bcrypt>=4.2,<5.0",
    "pyjwt>=2.10,<3.0",
    "python-multipart>=0.0.20,<1.0",
    "httpx>=0.28,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "ruff>=0.7,<1.0",
    "mypy>=1.13,<2.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
